Fix lack of support for dropped columns in foreign table. oracle_fdw incorrectly assumed that the n-th PostgreSQL column corresponds to the n-th Oracle column. This leads to confusion in the presence of dropped columns, which retain an entry in pg_attribute.
This patch adds entries "pgattnum" and "pgname" to struct oraColumn to track
names and attnum of the corresponding PostgreSQL columns. In the resulting
TupleTableSlot, dropped columns will be returned as NULLs.
"getOracleWhereClause", "getUsedColumns" and "oracleIterateForeignScan" are
changed to make use of the new "pgattnum".
While fixing it, rearrange "oracleIterateForeignScan" for improved readability.

While we're at it, split "checkDataTypes" in two functions:
- "getColumnData" which will populate the Oracle table description with
  data from the corresponding PostgreSQL table.
- "checkDataTypes", which does the actual check that the types match.

Bug found by Dominique Legendre.