Add join push-down feature.
*** NOTE ***
This patch requires Kohei KaiGai's custom/Foreign join patch(v9) which was
posted to pgsql-hackers on Mar 6 2015.
http://www.postgresql.org/message-id/9A28C8860F777E439AA12E8AEA7694F8010C140B@BPXM15GP.gisp.nec.co.jp
The code for join push-down is surrounded by #ifdef JOIN_API and #endif, so
oracle_fdw still can be compiled with old version of PostgreSQL.
*** NOTE ***

"Join push-down" is a feature that executes join between foreign tables
on remote side to reduce amount of data transfered from remote server,
and off-load join processing.  Its performance effect is outstanding in the
case that base tables are large but produced result is small.

A join to be pushed down must satisfy all conditions below:

1) The join is in SELECT statement.
UPDATE and DELETE can have join in source scan, but it requries extra
treatment, so in this first cut we don't support such joins.

2) It joins two foreign tables on the same server
This means that oracle_fdw supports push-down of 2-way join, so one of
"A JOIN B JOIN C" is executed on local side.  If we have a query contains
"A JOIN B JOIN C JOIN D", "A JOIN B" and "C JOIN D" (or some other
combination) would be executed on remote side.  This might be relaxed in
future.

3) Its join type must be INNER or one of OUTER joins.
CROSS join, a.k.a. certesian product is usually larger than total of
source relations, so we don't try to push-down such join.

4) No push-down-unsafe expression in the join
Any unsafe expression in join clause and other conditions supresses join
push-down.

3) Two foreign tables is accessed with same identity.
If one of join targets is behind a view owned by another user, access
check is done with the view owner.  We give up such joins because it might
result multiple user mappings.

For implementation issue, this commit disables optimization of SELECT
clause, listing only used column, but it would be able to enable that
optimization again in future.