Fix error with date/timestamp parameters
If a foreign table with a date/timestamp column is defined so that
the data types in Oracle and PostgreSQL are different (e.g. "date"
and "timestamp"), using a parameter for this column could fail with
"ORA-01856: BC/B.C. or AD/A.D. required".

This could happen with prepared statements or during data modification
if the column is part of the primary key.

This was an oversight in 76cb66bd6a9eef9222c96ed3a0e73a35611d1166
and used to work in version 1.2.0.

Found while investigating a bug report by "DFuster" (issue #98).