Fix crash when UPDATEing non-existent columns
If a foreign table has more columns than the underlying Oracle table,
these additional columns are treated as containing NULLs.
However, oracle_fdw crashed on an attempt to change such a column with
an UPDATE statement, which should not happen.

To fix, ignore such columns during UPDATE and emit an error message if
only such columns are updated, since we cannot issue an no-operation
UPDATE statement in Oracle and this is probably by mistake anyway.

This should fix #105 reported by Fabien Nicollet.