Clean up sort pushdown patch
Most of these changes only affect formating and placement of code.
There was a call to createQuery() with a wrong signature.

In passing, fix some compilation warnings with PostgreSQL 9.1 that
were not caused by the sort pushdown patch.

Also, remove the non-cosmetic changes to getOracleWhereClause()
because that doesn't seem quite right - this will be fixed with a
separate patch.