Fix bug in "pg_terminate_backend" handling
Oracle client overwrites the signal handlers, which causes
signals to be handled in ways different from how PostgreSQL
expects it.

While "pg_cancel_backend" (or SIGINT) seems to be handled
well (the running Oracle query is canceled and an error is
displayed), "pg_terminate_backend" (or SIGTERM) was not handled
correctly, since neither the Oracle query was canceled nor did
the backend terminate.

Fix by installing a signal handler for SIGTERM.

Reported by Dmitry Chirkin in issue #194.

In passing, improve the ordering of #includes.