Fix regression: No check for matching data types
In commit 330979f85bcaff7a78a02f266449173aaa590b10, createQuery
was moved to oracleGetForeignPlan, but the call to checkDataType
was left in oracleGetForeignRelSize.

This won't work, because checkDataType checks only the columns that are
used in the SELECT list, which are determined in createQuery.
So the check effectively didn't take place.

Fix by moving the check to oracleGetForeignPlan and adding a regression test.