Move PostGIS initialization from _PG_init to oracleGetSession
Before that, initializing background workers (e.g. for parallel query)
failed with
   ERROR:  invalid cache ID: 41
when oracle_fdw was loaded.

This is because catalog lookups need an active transaction, and there
is none when a background worker is set up.

This also fixes the known error you get when adding oracle_fdw to
shared_preload_libraries, which can now be done safely.

Closes #215.

Report by Chris Mair.