Enable more useful diagnostics on Windows.
By configuring oci.dll as delay-loaded, we can handle a failure loading it
and produce a better error message.

Without delay-loading, if oci.dll is not found, loading oracle_fdw itself
fails, and the default error message in this situation mentions only
oracle_fdw, causing the appearance that _it_ was not found rather than its
dependency.