Use correct SQLSTATE for some Oracle errors
We already passed on Oracle serialization failures as
serialization failures in PostgreSQL, and it seems like a
good idea to pass on constraint violation and deadlock
errors in a similar fashion.

This makes it easier for client programs to detect the cause
of the error; parsing the error message is unreliable.

Inspired by issue #234 from "omistler".