Make sure that only normal attributes are added to a join table
This should not happen normally, but before dece5310a3ee3f94a5e8bf775580eef34c74308d
a FOR UPDATE query would (for reasons unclear to me) try to add a
whole-row reference to the join table.

This not only caused spurious empty columns in the result, but also could
lead to a buffer overrun, since the code assumed that the number of columns
in the join table can be at most the sum of the number of columns in the
base tables.

Noted by Tatsuro Yamada.