Move oraTable creation to oracleGetForeignPlan for join relations
Creating it in oracleGetForeignJoinPaths was too early, because
the target list could still change.
This has not been a problem before PostgreSQL commit
11cf92f6e2e13c0a6e3f98be3e629e6bd90b74d5, but it was still wrong.

This caused crashes with pushed down joins with PostgreSQL v11.

The bug was tickled out by Tatsuro Yamada's regression tests.

The regression tests don't crash now, but fail because of the different
target lists in v11.
We will have to provide an alternative result file.