Fix push-down of conditions with variables referencing other tables
Apply the fix indicated in
https://www.postgresql.org/message-id/24678.1548870484%40sss.pgh.pa.us
to oracleGetForeignPaths.

The test case in that thread exposed another problem:
getUsedColumns failed to check if a variable in a scan condition
belonged to the foreign table being scanned or not.

Thanks to Tom Lane for the analysis and Tatsuro Yamada for the heads up.