Add support for reading infinite NUMBERs
Oracle NUMBER has infinite values, which are represented as "~" and "-~".
If you try to store such numbers in Oracle, it gags on them
(so much for robustness), but you can trick it into storing them,
and they occur in the field, so we'd better support them.

Since PostgreSQL's "numeric" does not know infinity, map these values
to NaN in this case.  For "real" and "double precision" we can use the
normal infinity values.

This patch adds no support for writing such values to Oracle.