Adapt to PostgreSQL commit 1cff1b95ab6ddae32faa3efe0d95a820dbfdc164
This commit changed the internal representation for lists from v13 on.

The access functions "lnext" and "for_each_list" now require a
third argument.

Add macros "list_next" and "do_each_list" that expand to the correct
definitions, based on the PostgreSQL version (it was not possible to
name the macros "lnext" and "for_each_list", as these are already
macros).

From now on, all code has to use "list_next" and "do_each_list" to
stay portable.