Fix regression tests
In the PostGIS regression test, an strange rounding error surfaced.
Fix by using a different coordinate that doesn't have the problem.

The join regression tests had an instability with respect to hash
and merge joins on PostgreSQL, so force it to use the former.

Also, remove the original join test expected file and keep only
the alternative file, since I don't want to bother making the
changes to both files.  We can always add an alternative result
file later if necessary.