Fix NCHAR and NVARCHAR2 handling
By default, Oracle converts NCHAR and NVARCHAR2 columns
(which are stored in a UNICODE encoding) to the database
encoding before converting them to the client encoding.

If the database encoding is not AL32UTF8 (which is the
main use case for the now obsolete NCHAR and NVARCHAR2
data types), converting to the database encoding will
substitute characters that cannot be represented with
"replacement characters", which results in data loss.

Reported by "srakazmus", who also found that you need
to use OCI_NCHAR_LITERAL_REPLACE_ON in OCIEnvCreate.

Special thanks to Christian Ullrich, who suspected that
OCI_ATTR_CHARSET_FORM needs to be set to SQLCS_NCHAR in
the bind and define handles to make this work.

Unfortunately the Oracle documentation is unclear and
contradictory on this point, so we had to use trial and
error.

This closes #416.