Fix performance regression introduced in 77242d1fef
Setting OCI_ATTR_CHARSET_FORM to SQLCS_NCHAR makes Oracle handle
encoding conversion for columns with a "national character set"
correctly, but it causes a severe performance impact, as reported
by Phil Florent in #434.

So improve the code to set the property only on define handles
for such Oracle columns.

The case of bind handles is left as it is - our DML performance
is not competitive anyway, and it would require tracking if a
parameter is bound to an Oracle column where it is needed.

This closes #434.