Improve the cookbook example with an explanation
The GRANT on the foreign server didn't have an explanation,
so many readers didn't understand that it was not required for
oracle_fdw to work, but was a recommendation for good security
practices.

As a consequence, many people who used oracle_fdw copied the
statement faithfully without understanding its purpose, even in
cases where it didn't have an effect (for example GRANT to the
superuser itself), and it has become an example of "cargo cult
programming".

This is clearly *not* what I intended, so add an explanation.

Per request from Tatsuro Yamada.