Fetch the Oracle server version during connect
Rather than fetching the server version when "oracleServerVersion"
is called, fetch it right after connecting to the Oracle server
and store it in the oracleSession.

That way, "oracleServerVersion" can be called without the overhead
of a client-server round trip.   This is useful for features that
want to check the Oracle server version, because they need to
behave differently for different Oracle server versions.