Push down LIMIT clauses if possible
These are pushed down as FETCH FIRST n ROWS ONLY from Oracle 12.2 on.
If OFFSET is present, it is added to the pushed down LIMIT.

Pushing down LIMIT is not strictly necessary, as PostgreSQL stops
fetching rows from Oracle as soon as the limit is reached anyway.
However, adding the limit to the Oracle query can give the Oracle
optimizer a clue, so that it can prefer a plan that returns the
first rows quickly.

Based on a patch by Gilles Darold, per request from Ilia Sazonov.

This closes #463 and also closes #465.