Use correct user mapping for views on foreign tables
Commit 31b0a51834 tried to fix that problem for SECURITY DEFINER
functions, but forgot the case of views on foreign tables, which
should use the user mapping of the owner of the view.

Revert 31b0a51834 and fix the problem by using the user ID
associated with the RangeTable for the foreign table
(the "checkAsUser" attribute) when it is set, otherwise the
current user ID.

This fixes #498 reported by "JSilex".