Use a different query hash than MD5
The problem is that calculating MD5 hashes fails on FIPS enabled
Linux systems, as detailed in
https://postgr.es/m/89B7F072-5BBE-4C92-903E-D83E865D9367%40trivadis.com

This caused problems in the past, since the signature of "pg_md5_hash"
changed, and it makes oracle_fdw unusable on FIPS enabled systems.

The solution is to use "hash_bytes_extended" instead, which also
simplifies the code.  Note that all the query hashes change, but since
they are only used internally anyway, that should be no problem.

This closes #519 by Hank Drews.