Fix processing for TIMESTAMP WITH LOCAL TIME ZONE
This Oracle data type behaves like PostgreSQL's "timestamp with time zone"
in that it is stored as absolute timestamp and converted to the Oracle
session ime zone on output.  But that output dies not have a time zone
offset, so PostgreSQL interperts it in its own session time zone, which
is wrong.

To fix, use AT TIME ZONE in Oracle when selecting such data.  Then the
output has a time zone offset, and PostgreSQL can handle it correctly.

Bug reported by "JamesInform".

This patch is lacking regression tests, so it is work in progress.
Let's first see if it fixes the reported problem.

This should fix #542.