Skip Oracle savepoints in read-only transactions
When the "isolation_level" is "read_only", there is no need to
manage subtransactions on the Oracle side.  This saves extra
round trips to Oracle.  In addition, there can be no Oracle
errors when a subtransaction ends, like in an exception handler
in PL/pgSQL code.

Per suggestion from "math-g".  This closes #558.