Don't use the deprecated OCIStmtPrepare
Rather, use OCIStmtPrepare2.  This requires a rewrite of allocHandle()
and freeHandle(), since now the statement handle has to be freed with
OCIStmtRelease().  Use an enum instead of a boolean for the handle kind
for more clarity and future extensibility.

Also, fix a bug in oracleEndTransaction(), introduced by b926831356:
we have to free all handles before we return, even if there is nothing
to commit.  Otherwise we are leaking memory.