Use the array interface instead of OCI prefetching
OCI prefetching is convenient, because it batches reads automatically,
but, like so many Oracle features, the implementation is incomplete:
it does not work if there are LOB, LONG or object columns.  This led to
bad performance when reading from Oracle, because every row required a
round trip.

Improve this by explicitly fetching arrays of rows.  For that, the
result buffer, the indicator and the result length have to be converted
to arrays.  Reuse the "prefetch" parameter for the size of this arrays.
The default value for the array size is reduced to 50.

The semantics of oracleExecuteQuery() and oracleFetchNext() have changed:
the former now returns the number of rows fetched or affected, while the
latter returns 0 if there is no more result and a 1-based offset into
the result array otherwise.

To speed up fetching of LOBs and BFILEs, set OCI_ATTR_LOBPREFETCH_SIZE
and OCI_ATTR_LOBPREFETCH_LENGTH on the respective define handles, so that
Oracle fetches the beginning and the length of the LOB with the row.
The LOB prefetch size is exposed as a new table option "lob_prefetch",
which defaults to 1MB.

I gave up on using array fetching for SDO_GEOMETRY objects for now,
since the complexity of changing that is high and can destabilize code
that I am not very familiar with.

While reworking and simplifying the code in acquireSampleRowsFunc(),
which was far too cute for its own good, fix a potential bug and memory
leak: I had forgotten to free the statement handle.

Development of this enhancement was generously sponsored by Mipih
(https://www.mipih.fr/).