Make auto_explain work with oracle_fdw
oracleExplain and oracleQueryPlan used the OCIStmt pointer stored in
the oracleSession, which works fine if EXPLAIN is called as an SQL statement.
But if auto_explain is used, oracleExplain gets called while the actual
statement has not been closed yet.
That leads to an error message:
  ERROR:  oracleQueryPlan internal error: statement handle is not NULL
The problem is fixed by using a separate OCIStmt pointer.

Reported by "Jaisus" and Sebastien Caunes.  Sebastien provided the essential
clue that the problem is related to auto_explain.

This closes #359.