Allocate enough memory for dummy result columns
If we need no column from the Oracle side (in queries like "SELECT 27 FROM ..."
or "SELECT count(*) FROM ..."), we use "SELECT 1 FROM ..." to construct a valid
Oracle SQL statement.

When commit d679a73d02 introduced array prefetching, I forgot to allocate
sufficient memory for that dummy result.  Since the result value for that dummy
column was defined as a static variable, that smashed the stack, leading to
mysterious crashes.

Thanks to Philippe Florent and "JamesInform" for the bug report.
Special thanks to Philippe for his tireless help in diagnosing the cause!

This closes #582.