Don't push down LIMIT without ORDER BY
As shown by issue #613, it is not always safe to push down LIMIT
if there is no ORDER BY to be pushed down.  This patch will cause
LIMIT to be pushed down in fewer cases (also where it would be safe),
but let's not worry about that too much, because LIMIT without
ORDER BY is pretty useless in real life anyway.

Per report from "fbattke".

This closes #613.