Free LOB locators when the statement ends
Commit 68b97fe927 made sure that all statement handles and
LOB locators got freed when the transaction ends.  This was
done by maintaining a list of pointers for these data structures
for each database connection.  Commit e8a3317e6c changed the
code, but the list stayed fundamentally the same.

The problem with this approach is that LOB locators are not
released before the end of the transaction.  If a transaction
runs many statements that return LOBs, these LOB locators
will accumulate.  This was exacerbated by commit d679a73d02,
since we now allocate a whole array of LOB locators.

Issue #621 shows how this can cause out-of-memory errors
and crashes (due to bugs in Oracle client).

This patch fixes the problem by maintaining a list of
statements for each connection and a list of LOB locators for
each statement.  When the statement is closed, all its LOB
locators are freed.

Report by "JosefMachytkaNetApp".