Turn import_* options into skip_*
As evidenced by my mistake in the regression tests in the previous
commit, having import_* options for IMPORT FOREIGN SCHEMA that default
to "true" is not intuitive, so let's negate them to skip_* options.
Users don't expect options to default to "true", and that way we cannot
forget any odd object types.