Fix query for IMPORT FOREIGN SCHEMA
Commits 4a012436be and 6b1e9663a8 didn't consider that there
micht be different objects with the same name in the same schema.
For example, there may be a table and an index with the same
name in the same schema.

That can lead to repeated results, which in turn leads to error
messages like

    ERROR:  column "id" specified more than once

To fix, change NOT IN to IN and explicitly enumerate the object
types that we care about.  This is safe, because tables, views
and materialized views in the same schema must have different names.

Per report from "JamesInform".

This closes #673.