Fix IMPORT FOREIGN SCHEMA for materialized views
Materialized views show up as TABLE *and* as MATERIALIZED VIEW
in the Oracle catalog view ALL_OBJECTS.  Fix the query by ignoring
the former if both are present.

Add a materialized view to the regression tests and improve the
coverage for IMPORT FOREIGN SCHEMA.

Per diligent testing by "JamesInform" in #673.