Update the check for unsupported minor versions
That check was added in d137d15edc.  Later, there were more ABI and API
breaks documented (commits 07e0b4562f and 33c4086ac3), but the check was
never updated.

In the case of an ABI break, execution will probably never reach that
check, because loading the oracle_fdw shared library fails.  Still, it
is better to be consistent.