Bind empty string parameters as NULL values
The attempt to bind a parameter as an empty string SQLT_LVB
leads to the error

  ORA-01459: invalid length for variable character string

This is particularly easy to trigger with an INSERT; inserting
an empty string is enough.

Work around the problem by binding a NULL value instead.
This usually does the right thing, because Oracle treats empty
strings as NULL values anyway.  The only case where this is not
true is with LOBs: they can be empty, but NOT NULL.
But getting that special case right would be a challenge, and
the fact that nobody has reported that problem until now is a
strong indication that few people will care.

Per report from Adrian Boangiu.
This fixes #679.