Increase "prefetch" maximum to 10240
This was the limit before commit b4e21374b5.
But both #671 and #674 request raising the limit because they
observe better performance with values over 1000, so let's
give in and raise the limit to 10240 again.

Since such large values are likely to offer no performance
benefit in most cases, and there is a high risk of
out-of-memory errors as seen in #582, add a dire warning to
the README.