Don't fetch more of a LOB than is needed
The change in commit d62b876149 inadvertently caused the whole
LOB to be fetched, even if we only need part of it (for ANALYZE).

Improve matters by exiting the loop as soon as we have enough.
This may still fetch the whole LOB from Oracle, but we can at
least only copy as much as we need.

Noticed by "shaunxzy" in #707.