Fix "nchar" bug introduced in 5fd556f602
That commit tried to reduce the impact of the "nchar" option
by only choosing the correct, expensive conversion for
NVARCHAR2 and NCHAR columns.  Unfortunately, an omission in
oracleDescribe() made oracle_fdw think that such columns
were in fact VARCHAR2 and CHAR, so the correct conversion
was never used.

This fixes #706 and (now for real) #416.

Report and analysis by Andrey Teplitskiy.